requirements.txt
google-cloud-datacatalog==3.25.1
google-cloud-bigquery==3.30.0